Move libraries into directories following package path
