+++ include/linuxdeploy/core/desktopfile/desktopfileentry.h
