+++ src/core/copyright/copyright.cpp
