+++ src/core/copyright/copyright.h
