+++ src/core/copyright/copyright_dpkgquery.cpp
