+++ src/core/copyright/copyright_dpkgquery.h
