+++ tests/core/test_desktopfilereader.cpp
@@ -3,7 +3,7 @@
 #include <boost/filesystem.hpp>
 
 // local headers
-#include "../../src/core/desktopfilereader.h"
+#include "../../src/core/desktopfile/desktopfilereader.h"
 
 using namespace linuxdeploy::core::desktopfile;
 namespace bf = boost::filesystem;
